moving UT to more appropriate class
