+++ azure-mgmt-network/src/test/resources/session-records/canDeleteNetworkWithServiceCallBack.json
